genai-doc-ingestion/requirements.txt
# FastAPI and related
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
pydantic-settings>=2.1.0

# AWS
boto3>=1.34.0
botocore>=1.34.0
opensearch-py>=2.4.2

# Document Processing
PyPDF2>=3.0.1
python-docx>=1.0.1
sentence-transformers>=2.5.0
nltk>=3.9.0

# External connectors - NOTE: pysnow requires pytz<2020.0
# If you need to use ServiceNow features, consider installing pytz==2019.3 locally
Office365-REST-Python-Client>=2.6.0  # SharePoint
# pysnow>=0.7.17  # ServiceNow - Commented out due to pytz conflict

# Alternative ServiceNow client options:
# 1. Fork pysnow and remove the restrictive dependency
# 2. Use requests directly to interact with ServiceNow API
# 3. Create a separate virtual environment for ServiceNow operations
textract-trp>=0.1.3  # AWS Textract helper

# ML and Vector processing
numpy>=1.26.0
scikit-learn>=1.4.0
spacy>=3.7.4
transformers>=4.40.0
torch>=2.2.0

# Frontend
streamlit>=1.32.0
plotly>=5.19.0
pandas>=2.2.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.2
regex>=2024.5.0
joblib>=1.3.2
pyyaml>=6.0.1
pytz>=2025.2  # Required by pandas

# Development
pytest>=7.4.4
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
httpx>=0.27.0 